remove un necessary widgets and refactor add workspace widgets
